requirements.txt
torch==1.7.1
torchvision==0.8.2
pytorch-lightning==1.1.8
Pillow==8.1.0
numpy~=1.20.1